Restructure directories and wire up data render fn
Moved header fns and styling to individual files
in component directory. Wired up data render
functionality to display a default location or
take input value and pass to API call. At this
stage of development, date and time fn calls users
local date and time from JSON, not specific to
the location the user is searching for.