Remove date/time and reposition location title
In this commit, I removed the date/time display from
the weather app and repositioned the location title.
However, this resulted in breaking the renderWeatherData function in index.js due to the getLocationFromUserInput function in header.js no longer being functional.

To address this issue, I plan to reroute the getLocationFromUserInput function call to the drop
down menu search feature in future updates. In the
meantime, I will fix the function to restore the app's functionality.

This commit represents a work in progress, but I
wanted to make sure to document the changes and the
issues that need to be resolved before merging with
main.