Add functionality for calling forecasted weather data API
In this commit, I added the functionality to call the "onecall" endpoint of the weather API that returns forecasted weather data for eight days. I also created a corresponding "daily" property that maps over the array of properties in the "renderWeatherData" function in "app.js" to display the forecasted weather in the forecast cards.

To accomplish this, I updated the "extractWeatherData" function in "weather-data.js" to include the "onecall" endpoint. Additionally, I created a new function named "getLocationCoords" that takes a location name in string form and returns its latitude and longitude, which are required for the "onecall" endpoint API call.

Lastly, I created a function that returns the API key instead of using a free-floating constant.