Fix bug causing default search on unit conversion click
This commit resolves a bug where the user's search location was being
replaced by the default location when the user clicked a temperature
unit conversion button. To fix this, I created a new object,
locationState, which tracks the user's search location throughout the
application.

When the user submits a new search location, formSubmissionHandler
updates the locationState object with the new location value. On each
renderWeatherData function call, the current location is retrieved
using the getLocation method, which uses the locationState object.

To simplify the writeLocation function in header.js, I added a new
location parameter. In renderWeatherData, the location is sent to
the writeLocation function as an argument, rather than retrieving the
location from the search bar value.

This commit ensures that the user's chosen location is preserved when
they interact with the temperature unit conversion buttons.